+++ src/layout/Modal/ModalVacancy/ModalVacancy.module.scss
@@ -8,6 +8,7 @@
 .main {
   display: flex;
   flex-direction: column;
+  gap: 10px;
 }
 
 .title {
